Separate typescript source files into their own directory.
Closes #196.
