+++ src/md5-hex.ts
