+++ src/strategies/persistent.ts
